pref：修复了部分table的展示效果

+++ src/views/PageContainer/ItemManage/ItemManageRegion/components/ManageRegions.js
@@ -34,26 +34,26 @@ export default function ManageRegions(props) {
 
     const tableColumns = [
         {
-            title: '规则编码',
+            title: '区划编码',
             dataIndex: 'region_code',
             key: 'region_code',
             width: 120
         },
         {
-            title: '规则路径',
+            title: '区划路径',
             dataIndex: 'region_path',
             key: 'region_path'
         },
         {
             title: '业务部门',
-            dataIndex: 'department',
-            key: 'department',
+            dataIndex: 'department_name',
+            key: 'department_name',
             width: 125
         },
         {
             title: '创建人',
-            dataIndex: 'creator',
-            key: 'creator',
+            dataIndex: 'creator_name',
+            key: 'creator_name',
             width: 100
         },
         {
@@ -190,31 +190,58 @@ export default function ManageRegions(props) {
         } 
         // 根据事项规则id删除事项规则，删除完之后重新载入事项规则
         api.DeleteRegions(data).then(response=>{ 
-            props.deleteRegionSimulate(deletingIds)
-            getRegions()
-            props.showSuccess()
+            if ('code' in response.data.data){
+                Modal.confirm({
+                    title: '规则已绑定',
+                    content: '所选部分规则已被部分事项绑定，若要删除则需要先进行解绑。是否跳转到事项流程管理？',
+                    centered: true,
+                    onOk: function(){
+                        let codes = []
+                        for (let i = 0; i < deletingIds.length; i++){
+                            codes.push(props.regionNodes[deletingIds[i]].region_code)
+                        }
+                        props.setBindedData({
+                            region_code: codes
+                        })
+                    }
+                })
+            }
+            else{
+                props.deleteRegionSimulate(deletingIds)
+                getRegions()
+                props.showSuccess()
+            }
+            
         }).catch(error=>{
             // 删除报错时，弹出报错框并重新加载数据
             props.showError('删除规则失败！')
             props.getRegionTree()
+            setCurrent(0)
         })
     }
 
+    useEffect(function(){
+        for (let key in props.bindedData){
+            props.jumpToProcess()
+            break
+        }  
+    }, [props.bindedData])
+
     const getRegions = ()=>{
         setTableLoading(true)
-        // 无搜索条件获取全数据
-        api.GetRegions({
-            page_num: 0,
-            page_size: currPageSize
-        }).then(response=>{
+        // 获取数据
+        let data = originData
+        data['page_num'] = current
+        data['page_size'] = currPageSize
+        api.GetRegions(data).then(response=>{
             let regions = response.data.data.data
             setTotalSize(response.data.data.total)
-            let table = []
             for (let i = 0; i < regions.length; i++){
+                regions[i]['department_name'] = regions[i].creator.department_name
+                regions[i]['creator_name'] = regions[i].creator.name
                 regions[i]['region_path'] = getPathByRegionId(regions[i]._id)
-                table.push(regions[i])
             }
-            setTableData(table)
+            setTableData(regions)
             setTableLoading(false)
         }).catch(error=>{
             props.showError('获取规则失败！')
@@ -233,12 +260,12 @@ export default function ManageRegions(props) {
             let regions = response.data.data.data
             setCurrent(0)
             setTotalSize(response.data.data.total)
-            let table = []
             for (let i = 0; i < regions.length; i++){
+                regions[i]['department_name'] = regions[i].creator.department_name
+                regions[i]['creator_name'] = regions[i].creator.name
                 regions[i]['region_path'] = getPathByRegionId(regions[i]._id)
-                table.push(regions[i])
             }
-            setTableData(table)
+            setTableData(regions)
             setTableLoading(false)
         }).catch(error=>{
             props.showError('搜索规则失败！')
@@ -263,7 +290,24 @@ export default function ManageRegions(props) {
         // 回 归 本 源
         setOriginData({})
         setCurrent(0)
-        getRegions()
+        setTableLoading(true)
+        api.GetRegions({
+            page_num: 0,
+            page_size: currPageSize
+        }).then(response=>{
+            let regions = response.data.data.data
+            setTotalSize(response.data.data.total)
+            for (let i = 0; i < regions.length; i++){
+                regions[i]['department_name'] = regions[i].creator.department_name
+                regions[i]['creator_name'] = regions[i].creator.name
+                regions[i]['region_path'] = getPathByRegionId(regions[i]._id)
+            }
+            setTableData(regions)
+            setTableLoading(false)
+        }).catch(error=>{
+            props.showError('重置失败！')
+            setTableLoading(false)
+        })
     }
 
     const changePage = (page, pageSize)=>{
@@ -281,6 +325,8 @@ export default function ManageRegions(props) {
             let regions = response.data.data.data
             let table = []
             for (let i = 0; i < regions.length; i++){
+                regions[i]['department_name'] = regions[i].creator.department_name
+                regions[i]['creator_name'] = regions[i].creator.name
                 regions[i]['region_path'] = getPathByRegionId(regions[i]._id)
                 table.push(regions[i])
             }
