+++ src/views/PageContainer/ItemManage/ItemManageRule/components/ManageRules.js
@@ -1,4 +1,5 @@
 import React, {cloneElement, useEffect, useState} from 'react'
+import { Redirect } from 'react-router-dom'
 import { DatePicker, Space, Dropdown, Menu, Button, Select, Table, Modal,Descriptions, Badge, message  } from 'antd';
 import { getYMD } from "../../../../../utils/TimeStamp";
 import api from '../../../../../api/rule';
@@ -42,14 +43,14 @@ export default function ManageRules(props) {
         },
         {
             title: '业务部门',
-            dataIndex: 'department',
-            key: 'department',
+            dataIndex: 'department_name',
+            key: 'department_name',
             width: 125
         },
         {
             title: '创建人',
-            dataIndex: 'creator',
-            key: 'creator',
+            dataIndex: 'creator_name',
+            key: 'creator_name',
             width: 100
         },
         {
@@ -183,9 +184,23 @@ export default function ManageRules(props) {
         } 
         // 根据事项规则id删除事项规则，删除完之后重新载入事项规则
         api.DeleteRules(data).then(response=>{ 
-            props.deleteRuleSimulate(deletingIds)
-            getRules()
-            props.showSuccess()
+            if ('code' in response.data.data){
+                Modal.confirm({
+                    title: '规则已绑定',
+                    content: '所选部分规则已被部分事项绑定，若要删除则需要先进行解绑。是否跳转到事项流程管理？',
+                    centered: true,
+                    onOk: function(){
+                        props.setBindedData({
+                            rule_id: deletingIds
+                        })
+                    }
+                })
+            }
+            else{
+                props.deleteRuleSimulate(deletingIds)
+                getRules()
+                props.showSuccess()
+            } 
         }).catch(error=>{
             // 删除报错时，弹出报错框并重新加载数据
             props.showError('删除规则失败！')
@@ -194,17 +209,24 @@ export default function ManageRules(props) {
         setDeletingIds([])
     }
 
+    useEffect(function(){
+        for (let key in props.bindedData){
+            props.jumpToProcess()
+            break
+        }  
+    }, [props.bindedData])
+
     const getRules = ()=>{
         // 无搜索条件获取全数据
         // 由于数据量较小，不进行分页处理，全部拉取
         api.GetRules({}).then(response=>{
             let rules = response.data.data
-            let table = []
             for (let i = 0; i < rules.length; i++){
+                rules[i]['department_name'] = rules[i].creator.department_name
+                rules[i]['creator_name'] = rules[i].creator.name
                 rules[i]['rule_path'] = getPathByRuleId(rules[i].rule_id)
-                table.push(rules[i])
             }
-            setTableData(table)
+            setTableData(rules)
         }).catch(error=>{
             props.showError('获取规则失败！')
         })
@@ -215,12 +237,12 @@ export default function ManageRules(props) {
         api.GetRules(data).then(response=>{
             let rules = response.data.data
             setCurrent(1)
-            let table = []
             for (let i = 0; i < rules.length; i++){
+                rules[i]['department_name'] = rules[i].creator.department_name
+                rules[i]['creator_name'] = rules[i].creator.name
                 rules[i]['rule_path'] = getPathByRuleId(rules[i].rule_id)
-                table.push(rules[i])
             }
-            setTableData(table)
+            setTableData(rules)
         }).catch(error=>{
             props.showError('搜索规则失败！')
         })
